requirements.txt
bctpy==0.6.1
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
communities==3.0.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.42.1
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lit==16.0.6
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
nibabel==5.1.0
nichord==0.2.7
nilearn==0.10.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
patsy==0.5.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-louvain==0.16
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.13.0
six==1.16.0
statsmodels==0.14.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
zipp==3.16.2
